Reorganize Book extension, use Haml programatically
Still a work in progress, but this large commit establishes what looks
like will be the best way forward.

At first I attempted to generate all XML programatically via Nokogiri,
in order to keep all epub-generation code confined to the extensions
directory. When it became apparent that this was going to be very
tedious and error-prone, I looked at using Haml for XML generation.

Using Haml has a lot of advantages (constructing a complex XML template
really is a task that needs a DSL, or else you are asking for trouble),
but my limited knowledege of Haml led me to think I'd need to put these
templates in the /layouts dir and create epub-specific resources.

It turns out there is a better way – Haml can be used programatically,
you just need to create a new Haml::Engine instance and feed it:

- A HAML template string
- any necessary options (a :format => :xhtml hash, in this case)

Then you can call Haml::Engine#render(Object.new, { :data => data })
where data is whatever object you are passing to the template context,
and :data is the name it has inside of that context locally. The output
is a properly-formatted HTML or XHTML doc.

See Book::Epub#load_template and Book::Epub#build_opf for an example of
this in action.

This allows the boilerplate XML layouts (Epub has so many...) to move
entirely inside of the extension; I'm leaving epub_chapter as a
contentional layout because this way the user can go in and make changes
if desired; no one needs to tamper with the template for the content.opf
or toc.ncx files, however.