Build out more boilerplate files
This commit adds the ability to build the boilerplate mimetype and
container.xml files programatically. Next steps are to repeat for the
other required components of a well-designed Epub (toc.ncx, nav file,
cover page, etc).

Note: Haml doesn't seem to like when a template is provided as a
heredoc inside a method (with the <<- delimiters); it throws errors
complaining about incorrect indentation. So everything longer than one
line will need to have its own Haml template file, it seems.