Valid Epub generation
Valid epub3 generation is now working. There are templates for most
essential features (toc page, toc.ncx file, cover page, basic css file,
etc) and the output is zipped up into a .epub file automatically with
the epzip gem. Users have the option of specifying a cover image if
desired.

There is a lot of boilerplate code here, and probably things could be
streamlined more, but this is an important milestone.