Explicitly require Nokogiri and /dist dir
This is a hedge against errors which seem to happen on some systems but
not others - on some systems nokogiri needs to be explicitly required
here even though it is also a dependency of Middleman.

Also, a .keep file has been forcibly included in /dist so that everyone
has that folder by default (epub process breaks without one). In the
future this directory should create itself it does not exist already.